Inherit common functions in home and standingStill
